refactor to generate local driving grid with a method (local_driving_grid) in frontend_methods.py
